ps1_bridge/requirements.txt
# RustChain PS1 Bridge - Python Dependencies

pyserial>=3.5
requests>=2.28.0
